Move hover tests into their own file
No code changes
